+++ review_apps/show.go
@@ -12,29 +12,34 @@ import (
 	"github.com/Scalingo/cli/utils"
 )
 
-func Show(app string) error {
+func Show(appID string) error {
 	c, err := config.ScalingoClient()
 	if err != nil {
 		return errgo.Notef(err, "fail to get Scalingo client")
 	}
 
-	reviewApps, err := c.SCMRepoLinkReviewApps(app)
+	reviewApps, err := c.SCMRepoLinkReviewApps(appID)
 	if err != nil {
 		return errgo.Notef(err, "fail to get review apps for this app")
 	}
 	if len(reviewApps) == 0 {
-		io.Statusf("No review app for '%s' or specified app is not a parent app.\n", app)
+		io.Statusf("No review app for '%s' or specified app is not a parent app.\n", appID)
 		return nil
 	}
 
 	t := tablewriter.NewWriter(os.Stdout)
-	t.SetHeader([]string{"App", "PR", "PR Branch", "Created At", "Status"})
+	t.SetHeader([]string{"App", "PR", "PR Branch", "Created At", "Status", "URL"})
 	for _, ra := range reviewApps {
 		date := ra.CreatedAt.Local().Format(utils.TimeFormat)
 
+		app, err := c.AppsShow(ra.AppID)
+		if err != nil {
+			return errgo.Notef(err, "fail to get app from review app")
+		}
+
 		t.Append([]string{
 			ra.AppName, fmt.Sprintf("%d", ra.PullRequest.Number), ra.PullRequest.BranchName,
-			date, fmt.Sprintf("%v", ra.LastDeployment.Status),
+			date, fmt.Sprintf("%v", ra.LastDeployment.Status), app.Url,
 		})
 	}
 	t.Render()